add test for capturing container exit code
